+++ gwcelery/conf/__init__.py
@@ -115,7 +115,7 @@ superevent_default_d_t_end = 1.0
 superevent_far_threshold = 1 / 3600
 """Maximum false alarm rate to consider events superevents."""
 
-significant_alert_far_threshold = {'cbc': 1 / (60 * 86400),
+significant_alert_far_threshold = {'cbc': 1 / (30 * 86400),
                                    'burst': 1 / (365 * 86400),
                                    'test': 1 / (30 * 86400)}
 """Group specific maximum false alarm rate to consider
@@ -131,6 +131,9 @@ early_warning_alert_trials_factor = 4.0
 """Trials factor for early warning alerts. There are two pipelines that are
 producing early warning events: gstlal and spiir."""
 
+early_warning_alert_far_threshold = 1 / (30 * 86400)
+"""False alarm rate threshold for significant early warning alerts."""
+
 significant_alert_trials_factor = dict(cbc=5.0, burst=4.0)
 """Trials factor corresponding to trigger categories. For CBC and Burst, trials
 factor is the number of pipelines. CBC pipelines are gstlal, pycbc, mbta and
@@ -138,9 +141,6 @@ spiir. Burst searches are cwb.allsky, cwb.bbh and cwb.imbh."""
 
 preliminary_alert_trials_factor = dict(cbc=8.0, burst=8.0)
 """Trials factor for less significant alerts."""
-
-early_warning_alert_far_threshold = 1 / (3600 * 24)
-"""False alarm rate threshold for early warning alerts."""
 
 snews_gw_far_threshold = 1 / (3600 * 24)
 """Maximum false alarm rate for a superevent to send out a coincidence alert